// frontend/src/shared/theme/types.ts
export type Theme = "dark" | "light" | "system";
export type Settings = {
  sidebarProgressBar: boolean;
  sidebarFlagsStats: boolean;
  mainPageCardOnly: boolean;
  mainPageNoMargins: boolean;
  mainPageFor4K: boolean;
  mainPageSlider: boolean;
  mainPageList: boolean;
  mainPageGrid: boolean;
  mainPageLarge: boolean;
  flagsVisibility: boolean;
};